Moved oauth test to secure_test.go
