+++ controllers/secure_test.go
@@ -1,0 +1,58 @@
+package controllers_test
+
+import (
+	"github.com/18F/cf-deck/controllers"
+	"github.com/18F/cf-deck/helpers"
+	. "github.com/18F/cf-deck/helpers/testhelpers"
+	"github.com/gocraft/web"
+	"golang.org/x/net/context"
+
+	"fmt"
+	"strings"
+	"testing"
+)
+
+var oauthTests = []basicAPITest{
+	{
+		BasicConsoleUnitTest: BasicConsoleUnitTest{
+			TestName:    "Basic Valid OAuth Session",
+			SessionData: ValidTokenData,
+		},
+		expectedResponse: "test",
+	},
+	{
+		BasicConsoleUnitTest: BasicConsoleUnitTest{
+			TestName:    "Basic Invalid OAuth Session",
+			SessionData: InvalidTokenData,
+		},
+		expectedResponse: "{\"status\": \"unauthorized\"}",
+	},
+}
+
+func TestOAuth(t *testing.T) {
+	mockSettings := helpers.Settings{}
+	mockSettings.TokenContext = context.TODO()
+
+	for _, test := range oauthTests {
+		// Initialize a new session store.
+		store := MockSessionStore{}
+		store.ResetSessionData(test.SessionData, "")
+		mockSettings.Sessions = store
+
+		// Setup a test route on the API router (which is guarded by OAuth)
+		response, request := NewTestRequest("GET", "/v2/test")
+		router := controllers.InitRouter(&mockSettings)
+		secureRouter := router.Subrouter(controllers.SecureContext{}, "/")
+		apiRouter := secureRouter.Subrouter(controllers.APIContext{}, "/v2")
+		apiRouter.Middleware((*controllers.APIContext).OAuth)
+		apiRouter.Get("/test", func(c *controllers.APIContext, rw web.ResponseWriter, r *web.Request) {
+			fmt.Fprintf(rw, "test")
+		})
+
+		// Make the request and check.
+		router.ServeHTTP(response, request)
+		if strings.TrimSpace(response.Body.String()) != test.expectedResponse {
+			t.Errorf("Test %s did not meet expected value. Expected %s. Found %s.\n", test.TestName, test.expectedResponse, response.Body.String())
+		}
+	}
+}
